requirements.txt
aiohttp==3.5.4
appdirs==1.4.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
Babel==2.7.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cliff==2.15.0
cmd2==0.9.16
colorama==0.4.1
cryptography==2.7
debtcollector==1.21.0
decorator==4.4.0
dogpile.cache==0.7.1
idna==2.8
iso8601==0.1.12
jmespath==0.9.4
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.0.2
keystoneauth1==3.16.0
msgpack==0.6.1
multidict==4.5.2
munch==2.3.2
netaddr==0.7.19
netifaces==0.10.9
openstacksdk==0.34.0
os-service-types==1.7.0
osc-lib==1.13.0
oslo.config==6.11.0
oslo.i18n==3.23.1
oslo.serialization==2.29.1
oslo.utils==3.41.0
pbr==5.4.2
prettytable==0.7.2
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyperclip==1.7.0
pyrsistent==0.15.4
python-cinderclient==4.2.1
python-glanceclient==2.16.0
python-keystoneclient==3.20.0
python-novaclient==14.2.0
python-openstackclient==3.19.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requestsexceptions==1.4.0
rfc3986==1.3.2
simplejson==3.16.0
six==1.12.0
stevedore==1.30.1
urllib3==1.25.3
warlock==1.3.3
wcwidth==0.1.7
wrapt==1.11.2
yarl==1.3.0
